+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ## Unreleased
+* Automatically hide Save button when form contains only display fields (no editable inputs)
 * Add dynamic expression evaluation support for formItem types 'label' and 'text' (similar to 'html')
 * Add support for title attribute in formItem type 'label'
 * Remove close/cancel button from edit dialogs (dialog can be closed using the X button in title bar)